drawing/requirements.txt
absl-py==2.3.1
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.0
cv==1.0.0
cycler==0.12.1
easyocr==1.7.2
filelock==3.19.1
flatbuffers==25.12.19
fonttools==4.60.2
fsspec==2025.10.0
ImageIO==2.37.2
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.9.4
mediapipe==0.10.9
mpmath==1.3.0
networkx==3.2.1
ninja==1.13.0
numpy==2.0.2
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
packaging==25.0
pillow==11.3.0
protobuf==3.20.3
pyclipper==1.3.0.post6
pycparser==2.23
pyparsing==3.3.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
six==1.17.0
sounddevice==0.5.3
sympy==1.14.0
tifffile==2024.8.30
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
zipp==3.23.0
